requirements.txt
opencv-python~=4.10.0.84
mediapipe~=0.10.18
numpy~=1.26.4
cvzone~=1.6.1
torch
torchvision
timm
streamlit
Pillow
pyttsx3